refactor: split out utility functions
